Make code a bit nicer
